requirements.txt
finnhub-python==2.4.19
sendgrid==6.11.0
python-dotenv==1.0.0
requests==2.31.0
